move the inner class FileFormats from Library and Dataset models to enums/file_formats and update all the references
