+++ api/scpca_portal/enums/file_formats.py
@@ -1,0 +1,8 @@
+class FileFormats:
+    ANN_DATA = "ANN_DATA"
+    SINGLE_CELL_EXPERIMENT = "SINGLE_CELL_EXPERIMENT"
+
+    CHOICES = (
+        (ANN_DATA, "AnnData"),
+        (SINGLE_CELL_EXPERIMENT, "Single cell experiment"),
+    )
