+++ api/scpca_portal/models/dataset.py
@@ -2,6 +2,7 @@ import uuid
 
 from django.db import models
 
+from scpca_portal.enums.file_formats import FileFormats
 from scpca_portal.models import APIToken, ComputedFile
 from scpca_portal.models.base import TimestampedModel
 
@@ -11,15 +12,6 @@ class Dataset(TimestampedModel):
         db_table = "datasets"
         get_latest_by = "updated_at"
         ordering = ["updated_at"]
-
-    class FileFormats:
-        ANN_DATA = "ANN_DATA"
-        SINGLE_CELL_EXPERIMENT = "SINGLE_CELL_EXPERIMENT"
-
-        CHOICES = (
-            (ANN_DATA, "AnnData"),
-            (SINGLE_CELL_EXPERIMENT, "Single cell experiment"),
-        )
 
     id = models.UUIDField(primary_key=True, default=uuid.uuid4, editable=False)
